// src/app/core/models/designer-form-disposition/designer-form-disposition.model.ts
export class DesignerFormDispositionModel {
    _id: string;
    fieldtype: string;
    fieldname: string;
    displayname : string;
    required: boolean;
    lookupdata: object[];
    defaultvalue: string;
    form : {};
    lookupfieldid: string;
    formorder: Number;
    value: any;
    editable: boolean;
    status: string;
}